requirements.txt
torch==2.1.2
transformers==4.38.2
jupyter==1.0.0
ipywidgets==8.1.1
jsonschema==4.20.0
langchain
python-dotenv==1.0.0
langchain-openai
pandas
docx2txt
python-docx==1.1.2
accelerate==0.27.2